+++ src/groups/active.ts
@@ -1,12 +1,12 @@
 import * as vscode from 'vscode';
-import { Cache } from './basecache';
-import {handleActiveFileDirectoryChange, TS_LANGUAGE_ID, ExtensionState, setExtensionState, updateStatusBar} from '../extension';
-import {uriToContainingUri} from '../uritools';
-import {getPackageSourceUris} from '../targettools';
+import { Cache } from '../util/basecache';
+import {TS_LANGUAGE_ID, ExtensionState, setExtensionState, updateStatusBar} from '../extension';
+import {uriToContainingUri} from '../util/uritools';
 import { ActiveFile } from '../model/activeFile';
 import {MAX_PKG_SIZE, showDismissableMessage, showErrorMessage, updateMaxPackageSize} from '../userinteraction';
 import * as path from 'path';
-import { uriToBuild } from './filepathtools';
+import { uriToBuild } from '../util/filepathtools';
+import { getPackageSourceUris, packageTooLarge } from '../util/packagetools';
 
 const CHANGE_PACKAGE_LIMIT_BUTTON = 'Change max package size';
 const CACHE_SIZE: number = Number(vscode.workspace.getConfiguration('bazel-import').maxCacheSize);
@@ -20,31 +20,6 @@ export interface ActiveData {
     packageSources: Set<string>, // Uri string representation
     buildUri: vscode.Uri;
 }
-
-function urisFromStrings(strings: string[]): vscode.Uri[] {
-    return strings.map(strUri => vscode.Uri.parse(strUri)); 
-}
-
-async function loadPackageSources(document: vscode.TextDocument) {
-    const currentTargetPair = await getPackageSourceUris(uriToContainingUri(document.uri));
-    if (currentTargetPair === undefined) {
-        showErrorMessage(
-            `${path.basename(document.uri.fsPath)} failed to open`
-        );
-        return;
-    }
-    const packageSources = currentTargetPair[0]; 
-    const currentTarget = currentTargetPair[1];
-    const currBuildURI = currentTargetPair[2];
-
-    ActiveFile.data = {
-        packageSources: packageSources,
-        target: currentTarget,
-        buildUri: currBuildURI, 
-        documentState: document.getText(),
-        uri: document.uri
-    };
-};
 
 export async function updateActiveEditor(editor: vscode.TextEditor | undefined) {
     if (editor === undefined || !vscode.workspace.getConfiguration('bazel-import').enableDeletion) {
@@ -70,7 +45,6 @@ export async function updateActiveEditor(editor: vscode.TextEditor | undefined) 
         );
         return;
     }
-
 
     updateStatusBar(
         `Loading packages for deletion analysis. Current package size is ${vscode.workspace.getConfiguration('bazel-import').maxPackageSize}`,
@@ -121,13 +95,32 @@ export async function updateActiveEditor(editor: vscode.TextEditor | undefined) 
     setExtensionState(ExtensionState.active);
 }
 
+/** 
+ * Checks if the new document is in the old document's directory (i.e., does not need package update)
+*/
+function handleActiveFileDirectoryChange(document: vscode.TextDocument) {
+    let oldDir = undefined; 
+    if (ActiveFile.data) {
+        oldDir = uriToContainingUri(ActiveFile.data.buildUri);
+    }
+
+    let newDir = uriToContainingUri(document.uri); 
+    if (newDir.toString() === oldDir?.toString()) {
+        console.log("No need to reload the build packages");
+        ActiveFile.data.uri = document.uri;
+        ActiveFile.data.documentState = document.getText(); 
+        return undefined;
+    }
+    return newDir; 
+}
+
 /**
  * Sets the deletion enabled flag depending on the size of the current package and informs user of status
  */ 
 function validatePackageSize() {
     let deletion = 'enabled';
         
-    if ((ActiveFile.data.packageSources.length ?? MAX_PKG_SIZE + 1) > MAX_PKG_SIZE) {
+    if (packageTooLarge()) {
         vscode.window
             .showWarningMessage(
                 `${ActiveFile.data.packageSources.length} file(s) in package. Increase max package size? (current max: ${MAX_PKG_SIZE})`, 
@@ -144,9 +137,30 @@ function validatePackageSize() {
     showDismissableMessage(msg);
 }
 
-export function packageTooLarge(): boolean {
-    return (ActiveFile.data.packageSources.length ?? MAX_PKG_SIZE + 1) > MAX_PKG_SIZE;
+function urisFromStrings(strings: string[]): vscode.Uri[] {
+    return strings.map(strUri => vscode.Uri.parse(strUri)); 
 }
+
+async function loadPackageSources(document: vscode.TextDocument) {
+    const currentTargetPair = await getPackageSourceUris(uriToContainingUri(document.uri));
+    if (currentTargetPair === undefined) {
+        showErrorMessage(
+            `${path.basename(document.uri.fsPath)} failed to open`
+        );
+        return;
+    }
+    const packageSources = currentTargetPair[0]; 
+    const currentTarget = currentTargetPair[1];
+    const currBuildURI = currentTargetPair[2];
+
+    ActiveFile.data = {
+        packageSources: packageSources,
+        target: currentTarget,
+        buildUri: currBuildURI, 
+        documentState: document.getText(),
+        uri: document.uri
+    };
+};
 
 function setDeletionStatus(fileName: string) {
     const enabledStatus = packageTooLarge() ? "disabled" : "enabled";
