+++ src/util/exectools.ts
@@ -25,6 +25,7 @@ export async function executeCommand(command: string, cwd: string): Promise<stri
     return new Promise((resolve, reject) => {
         exec(command, {cwd: cwd}, (error, stdout, stderr) => {
             if (error) {
+                console.debug(error, stderr);
                 return reject({
                     "error": error,
                     "stderr": stderr
